refactor: 이동 정책 인터페이스 및 구현체 네이밍 정리
