+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import racingcar.policy.RandomValueMovePolicy;
+import racingcar.policy.MovePolicy;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class RacingGameTest {
@@ -14,7 +14,7 @@ class RacingGameTest {
     @Test
     void race_모든_자동차가_정상적으로_경주를_완료한다() {
         Cars cars = new Cars(List.of("자동차하나", "자동차둘"));
-        RandomValueMovePolicy movePolicy = () -> true;
+        MovePolicy movePolicy = () -> true;
         RacingGame game = new RacingGame(cars, movePolicy);
 
         RaceHistory history = game.race(2);
